catalogdb/requirements.txt
pymongo
faker
pandas         
pytest  
python-dotenv      
